[IMP] kicker: odoo13 javascript
Switch to PublicRoot for the root application, this means that we
now have a service provider and can use this._rpc for our remote calls.

Switch from jquery mobile to swiped-events for swiping event detection
for the menu since jquery mobile is not compatible with jquery3.

Use native promises instead of jquery promises/deffereds.